fixing a bunch of warnings, turning up the Release mode optimization level, and adding a vcproj.user file to make Lua script reloading not insanely slow by default when running from the debugger in Release mode (if you've opened the project before, you'll have to delete your snes9xw.vcproj.COMPUTERNAME.USERNAME.user file for this last change to take effect)

+++ cpuexec.cpp
@@ -378,7 +378,7 @@ void S9xDoHBlankProcessing ()
 	if (CPU.V_Counter >= FIRST_VISIBLE_LINE &&
 	    CPU.V_Counter < PPU.ScreenHeight + FIRST_VISIBLE_LINE)
 	{
-	    RenderLine (CPU.V_Counter - FIRST_VISIBLE_LINE);
+	    RenderLine ((uint8)(CPU.V_Counter - FIRST_VISIBLE_LINE));
 	}
 	
 	if(Settings.UseWIPAPUTiming)
